ajuste no back, data de requisicao e melhoria nos filtros do front

+++ backend/app/schemas/chamado.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel
@@ -31,7 +30,7 @@ class ChamadoBase(BaseModel):
     dataRequisicao: Optional[datetime] = None
     dataConclusao: Optional[datetime] = None
 
-    valorAprovado: Optional[Decimal] = None
+    valorAprovado: Optional[float] = None
 
 
 class ChamadoResponse(ChamadoBase):
